pyproject.toml
[project]
name = "agentic-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.1.1",
    "crewai>=1.6.1",
    "crewai-tools>=1.6.1",
    "faiss-cpu>=1.13.0",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "litellm>=1.80.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "tavily-python>=0.7.13",
]
